+++ karma.conf.js
@@ -4,10 +4,12 @@ module.exports = function(config) {
         // base path, that will be used to resolve files and exclude
         basePath: '',
 
-        preprocessors: {},
+        preprocessors: {
+            'tests/browserify.spec.js': ['browserify']
+        },
 
         // frameworks to use
-        frameworks: ['mocha', 'chai', 'sinon'],
+        frameworks: ['mocha', 'chai', 'sinon', 'browserify'],
 
 
         // list of files / patterns to load in the browser