+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+#### 0.2.0
+
+* CommonJS / Browserify support
+* Fixes bug where extended models don't inherit computed properties
+
 #### 0.0.3
 
 Fixes bug where a listener was setup even if there were no dependent properties